Inline methods for userCredentialManager in legacy code, moving LegacySessionSupportProvider to the same package where KeycloakSession resides in but server-spi-private, so that infinispan doesn't need to depend on services module
