+++ model/legacy-services/src/main/java/org/keycloak/services/legacysessionsupport/LegacySessionSupportProviderImpl.java
@@ -3,6 +3,7 @@ package org.keycloak.services.legacysessionsupport;
 import org.keycloak.credential.UserCredentialStoreManager;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.UserCredentialManager;
+import org.keycloak.models.LegacySessionSupportProvider;
 
 /**
  * @author Alexander Schwartz
