+++ services/src/main/java/org/keycloak/services/DefaultKeycloakSession.java
@@ -45,7 +45,7 @@ import org.keycloak.provider.ProviderFactory;
 import org.keycloak.provider.InvalidationHandler.InvalidableObjectType;
 import org.keycloak.provider.InvalidationHandler.ObjectType;
 import org.keycloak.services.clientpolicy.ClientPolicyManager;
-import org.keycloak.services.legacysessionsupport.LegacySessionSupportProvider;
+import org.keycloak.models.LegacySessionSupportProvider;
 import org.keycloak.sessions.AuthenticationSessionProvider;
 import org.keycloak.storage.DatastoreProvider;
 import org.keycloak.storage.federated.UserFederatedStorageProvider;
